First pass on GPU Particle System Fix FPS Locking RHI Buffer Refactor

+++ GraphicsEngine/Source/Core/Core/Platform/Windows/WindowsApplication.h
@@ -14,7 +14,7 @@ public:
 	static int ExecuteHostScript(std::string Filename, std::string Args, std::string WorkingDir, bool ShowOutput = false);
 	static void InitTiming();
 	static double Seconds();
-	static void Sleep(float Milliseconds);
+	static void Sleep(double Milliseconds);
 	static CORE_API void DisplayMessageBox(std::string title, std::string message);
 	static bool DisplayOpenFileDialog(std::string StartDir, std::string Filter, std::string & outData);
 	static bool DisplaySaveFileDialog(std::string StartDir, std::string Filter, std::string Extenstion, std::string & outData);
